+++ tests/testthat.R
@@ -10,6 +10,11 @@ mm8_db <- dbConnect(SQLite(), tempfile())
 createMsqlBackendDatabase(mm8_db, mm8_file)
 mm8_be <- backendInitialize(MsqlBackend(), mm8_db)
 
+mm14_file <- system.file("microtofq", "MM14.mzML", package = "msdata")
+mm_db <- dbConnect(SQLite(), tempfile())
+createMsqlBackendDatabase(mm_db, c(mm8_file, mm14_file))
+mm_be <- backendInitialize(MsqlBackend(), mm_db)
+
 tmt_file <- proteomics(full.names = TRUE)[4L]
 tmt_sps <- Spectra(tmt_file)
 tmt_db <- dbConnect(SQLite(), tempfile())