+++ tests/testthat/test_MsqlBackend.R
@@ -159,3 +159,14 @@ test_that("filterRt,MsqlBackend works", {
     expect_true(all(rtime(res2_1) >= 205 & rtime(res2_1) <= 210))
     expect_equal(filterMsLevel(res2, 2), filterMsLevel(res, 2))
 })
+
+test_that("filterDataOrigin works", {
+    res <- filterDataOrigin(mm_be, mm8_file)
+    expect_true(all(res$dataOrigin == mm8_file))
+
+    res <- filterDataOrigin(mm_be, mm14_file)
+    expect_true(all(res$dataOrigin == mm14_file))
+
+    res <- filterDataOrigin(mm_be, c(mm14_file, mm8_file))
+    expect_equal(unique(dataOrigin(res)), c(mm14_file, mm8_file))
+})